# docs-site/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Overlay API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/entitlements": {
      "get": {
        "operationId": "getEntitlements",
        "tags": [
          "entitlements"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/portal": {
      "post": {
        "operationId": "postPortal",
        "tags": [
          "portal"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription": {
      "get": {
        "operationId": "getSubscription",
        "tags": [
          "subscription"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "EntitlementsRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "EntitlementsResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
      "PortalRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "PortalResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
      "SubscriptionRequest": {
        "type": "object",
        "additionalProperties": true
      },
      "SubscriptionResponse": {
        "type": "object",
        "additionalProperties": true
      },
...
    }
  }
}