# src/app/api/portal/route.ts
import { NextRequest, NextResponse } from 'next/server'
import { stripe, getBaseUrl } from '@/lib/stripe'
import { getSession } from '@/lib/workos-auth'
import { convex } from '@/lib/convex'
import { resolvePortalConfigurationId } from '@/lib/stripe-billing'
import { getInternalApiSecret } from '@/lib/internal-api-secret'
import { enforceRateLimits, getClientIp } from '@/lib/rate-limit'

import { z } from '@/lib/api-schemas'

const PortalRequestSchema = z.object({ sessionId: z.string().optional() }).passthrough().openapi('PortalRequest')
const PortalResponseSchema = z.unknown().openapi('PortalResponse')
void PortalRequestSchema
void PortalResponseSchema

interface Subscription {
  userId: string
  email?: string
  stripeCustomerId?: string
  stripeSubscriptionId?: string
  tier: 'free' | 'pro' | 'max'
  planKind?: 'free' | 'paid'
  planAmountCents?: number
  status: 'active' | 'canceled' | 'past_due' | 'trialing'
}

async function findCustomerIdByEmail(email?: string): Promise<string | undefined> {
  if (!email) return undefined

  const customers = await stripe.customers.list({
    email,
    limit: 10,
  })

  if (customers.data.length === 0) return undefined

  const customersWithActiveSubscriptions = await Promise.all(
    customers.data.map(async (customer) => {
      const subscriptions = await stripe.subscriptions.list({
        customer: customer.id,
        status: 'all',
        limit: 10,
      })

      const activeSubscription = subscriptions.data.find((subscription) =>
        ['active', 'trialing', 'past_due'].includes(subscription.status),
      )

      return {
        customerId: customer.id,
        activeSubscription,
        created: customer.created ?? 0,
      }
    }),
  )

  const preferred =
    customersWithActiveSubscriptions.find((candidate) => candidate.activeSubscription)?.customerId ??
    customersWithActiveSubscriptions.sort((a, b) => b.created - a.created)[0]?.customerId

  return preferred
}

async function resolveExistingCustomerId(customerId?: string): Promise<string | undefined> {
  if (!customerId) return undefined

  try {
    const customer = await stripe.customers.retrieve(customerId)
    if (typeof customer === 'string') return customer
    if ('deleted' in customer && customer.deleted) return undefined
    return customer.id
  } catch {
    return undefined
  }
}

async function resolveVerifiedCustomerIdFromCheckoutSession(
  sessionId: string,
  userId: string,
): Promise<string | undefined> {
  const checkoutSession = await stripe.checkout.sessions.retrieve(sessionId)
  if (checkoutSession.metadata?.userId !== userId) {
    return undefined
  }
  return await resolveExistingCustomerId(checkoutSession.customer as string)
}

export async function POST(request: NextRequest) {
  try {
    const authSession = await getSession()
    if (!authSession || !authSession.user) {
      return NextResponse.json({ error: 'Authentication required' }, { status: 401 })
    }

    const userId = authSession.user.id
    const rateLimitResponse = await enforceRateLimits(request, [
      { bucket: 'billing:portal:ip', key: getClientIp(request), limit: 10, windowMs: 10 * 60_000 },
      { bucket: 'billing:portal:user', key: userId, limit: 5, windowMs: 10 * 60_000 },
    ])
    if (rateLimitResponse) return rateLimitResponse

    const body = await request.json().catch(() => ({}))
    const { sessionId } = body

    let customerId: string | undefined
    let subscriptionId: string | undefined

    // If we have a checkout session ID, get customer from there
    if (sessionId) {
      customerId = await resolveVerifiedCustomerIdFromCheckoutSession(sessionId, userId)
      if (!customerId) {
        return NextResponse.json(
          { error: 'Checkout session does not belong to the authenticated user.' },
          { status: 403 }
        )
      }
    }

    // Otherwise, look up customer from our database
    if (!customerId) {
      const subscription = await convex.query<Subscription>('subscriptions:getByUserId', {
        accessToken: authSession.accessToken,
        userId,
      })
      customerId = await resolveExistingCustomerId(subscription?.stripeCustomerId || undefined)
      subscriptionId = subscription?.stripeSubscriptionId || undefined

      if (!customerId && subscriptionId) {
        const stripeSubscription = await stripe.subscriptions.retrieve(subscriptionId)
        customerId = await resolveExistingCustomerId(
          typeof stripeSubscription.customer === 'string'
            ? stripeSubscription.customer
            : stripeSubscription.customer?.id,
        )
      }

      if (!customerId) {
        customerId = await findCustomerIdByEmail(authSession.user.email || subscription?.email)
      }

      if (customerId && !subscription?.stripeCustomerId) {
        await convex.mutation('subscriptions:upsertSubscription', {
          serverSecret: getInternalApiSecret(),
          userId,
          email: authSession.user.email,
          stripeCustomerId: customerId,
          stripeSubscriptionId: subscriptionId,
          tier: subscription?.tier,
          planKind: subscription?.planKind,
          planAmountCents: subscription?.planAmountCents,
          status: subscription?.status,
        })
      }
    }

    if (!customerId) {
      return NextResponse.json(
        { error: 'No customer found. Please subscribe first.' },
        { status: 400 }
      )
    }

    const baseUrl = getBaseUrl()

    // Create billing portal session
    const portalConfigurationId = resolvePortalConfigurationId()

    let portalSession
    try {
      portalSession = await stripe.billingPortal.sessions.create({
        customer: customerId,
        return_url: `${baseUrl}/account`,
        ...(portalConfigurationId ? { configuration: portalConfigurationId } : {}),
      })
    } catch (error) {
      const stripeError = error as { code?: string; message?: string; type?: string }
      const invalidConfiguration =
        stripeError?.code === 'resource_missing' ||
        (stripeError?.type === 'StripeInvalidRequestError' &&
          stripeError?.message?.toLowerCase().includes('configuration'))

      if (!portalConfigurationId || !invalidConfiguration) {
        throw error
      }

      console.warn(
        `[portal] Falling back to default portal configuration because ${portalConfigurationId} is invalid.`,
      )
      portalSession = await stripe.billingPortal.sessions.create({
        customer: customerId,
        return_url: `${baseUrl}/account`,
      })
    }

    return NextResponse.json({ url: portalSession.url })
  } catch (error) {
    console.error('Portal error:', error)
    return NextResponse.json(
      { error: 'Failed to create portal session' },
      { status: 500 }
    )
  }
}